+++ src/components/Button/Button.tsx
@@ -1,4 +1,4 @@
-// Tremor Button [v0.2.0]
+// Tremor Button [v1.0.0]
 
 import React from "react"
 import { Slot } from "@radix-ui/react-slot"
@@ -11,7 +11,7 @@ import { focusRing } from "../../utils/focusRing"
 const buttonVariants = tv({
   base: [
     // base
-    "relative inline-flex items-center justify-center whitespace-nowrap rounded-md border px-3 py-2 text-center text-sm font-medium shadow-sm transition-all duration-100 ease-in-out",
+    "relative inline-flex items-center justify-center whitespace-nowrap rounded-md border px-3 py-2 text-center text-sm font-medium shadow-xs transition-all duration-100 ease-in-out",
     // disabled
     "disabled:pointer-events-none disabled:shadow-none",
     // focus
@@ -30,7 +30,7 @@ const buttonVariants = tv({
         "hover:bg-blue-600 dark:hover:bg-blue-600",
         // disabled
         "disabled:bg-blue-300 disabled:text-white",
-        "disabled:dark:bg-blue-800 disabled:dark:text-blue-400",
+        "dark:disabled:bg-blue-800 dark:disabled:text-blue-400",
       ],
       secondary: [
         // border
@@ -43,7 +43,7 @@ const buttonVariants = tv({
         "hover:bg-gray-50 dark:hover:bg-gray-900/60",
         // disabled
         "disabled:text-gray-400",
-        "disabled:dark:text-gray-600",
+        "dark:disabled:text-gray-600",
       ],
       light: [
         // base
@@ -58,7 +58,7 @@ const buttonVariants = tv({
         "hover:bg-gray-300/70 dark:hover:bg-gray-800/80",
         // disabled
         "disabled:bg-gray-100 disabled:text-gray-400",
-        "disabled:dark:bg-gray-800 disabled:dark:text-gray-600",
+        "dark:disabled:bg-gray-800 dark:disabled:text-gray-600",
       ],
       ghost: [
         // base
@@ -71,7 +71,7 @@ const buttonVariants = tv({
         "bg-transparent hover:bg-gray-100 dark:hover:bg-gray-800/80",
         // disabled
         "disabled:text-gray-400",
-        "disabled:dark:text-gray-600",
+        "dark:disabled:text-gray-600",
       ],
       destructive: [
         // text color
@@ -84,7 +84,7 @@ const buttonVariants = tv({
         "hover:bg-red-700 dark:hover:bg-red-600",
         // disabled
         "disabled:bg-red-300 disabled:text-white",
-        "disabled:dark:bg-red-950 disabled:dark:text-red-400",
+        "dark:disabled:bg-red-950 dark:disabled:text-red-400",
       ],
     },
   },