+++ src/components/Dialog/Dialog.tsx
@@ -1,4 +1,4 @@
-// Tremor Dialog [v0.0.1]
+// Tremor Dialog [v1.0.0]
 
 import React from "react"
 import * as DialogPrimitives from "@radix-ui/react-dialog"
@@ -38,7 +38,7 @@ const DialogOverlay = React.forwardRef<
         // background color
         "bg-black/70",
         // transition
-        "data-[state=open]:animate-dialogOverlayShow",
+        "data-[state=open]:animate-dialog-overlay-show",
         className,
       )}
       {...props}
@@ -65,7 +65,7 @@ const DialogContent = React.forwardRef<
             // background color
             "bg-white dark:bg-[#090E1A]",
             // transition
-            "data-[state=open]:animate-dialogContentShow",
+            "data-[state=open]:animate-dialog-content-show",
             focusRing,
             className,
           )}