+++ src/components/Popover/Popover.tsx
@@ -81,7 +81,7 @@ const PopoverContent = React.forwardRef<
             // transition
             "will-change-[transform,opacity]",
             "data-[state=closed]:animate-hide",
-            "data-[state=open]:data-[side=bottom]:animate-slideDownAndFade data-[state=open]:data-[side=left]:animate-slideLeftAndFade data-[state=open]:data-[side=right]:animate-slideRightAndFade data-[state=open]:data-[side=top]:animate-slideUpAndFade",
+            "data-[state=open]:data-[side=bottom]:animate-slide-down-and-fade data-[state=open]:data-[side=left]:animate-slide-left-and-fade data-[state=open]:data-[side=right]:animate-slide-right-and-fade data-[state=open]:data-[side=top]:animate-slide-up-and-fade",
 
             className,
           )}