+++ src/hooks/useOnWindowResize.ts
@@ -1,8 +1,8 @@
-// Tremor useOnWindowResize [v0.0.0]
+// Tremor useOnWindowResize [v0.0.2]
 
 import * as React from "react"
 
-export const useOnWindowResize = (handler: { (): void }) => {
+export const useOnWindowResize = (handler: () => void) => {
   React.useEffect(() => {
     const handleResize = () => {
       handler()