+++ src/components/Badge/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Badge Changelog
 
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
 ## 0.0.1
 
 ### Changes