+++ src/components/BarChart/changelog.md
@@ -1,5 +1,11 @@
 # Tremor BarChart Changelog
 
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
 ## 0.2.1
 
 ### Changes