+++ src/components/Calendar/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Calendar Changelog
 
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
 ## 0.1.0
 
 ### Changes