+++ src/components/Divider/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Divider Changelog
 
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
 ## 0.0.2
 
 ### Changes