+++ src/components/Toggle/changelog.md
@@ -1,1 +1,13 @@
 # Tremor Toggle Changelog
+
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
+## 0.0.0
+
+### Changes
+
+- Initial release