+++ src/components/Tracker/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Tracker Changelog
 
+## 1.0.0
+
+### Changes
+
+- BREAKING CHANGE: Tailwind CSS v4
+
 ## 0.1.3
 
 ### Changes