+++ src/hooks/changelog.md
@@ -1,5 +1,11 @@
 # Tremor hooks Changelog
 
+## 0.0.2
+
+### Changes
+
+- Fix: Use a function type instead of a call signature.
+
 ## 0.0.1
 
 ### Changes